requirements.txt
numpy==2.0.0
openpyxl==3.1.5
pandas==2.2.2
psutil==6.0.0
pypyodbc==1.3.6
pytest==8.2.2
PyYAML==6.0.1
PyYAML==6.0.1
SQLAlchemy==2.0.30
